More work on the theme.
Successfully got menus working as well as worked on the posts page.

The homepage (homepage.php) will be what we use as our index page (assigning it to the index page needs to be done in wordpress settings.

To have a post appear on the homepage, it needs to be in the homepage category. Additionally, it needs to be tagged with one of full, large, medium or small.