Added support to the theme for wordpress to use html5 (functions.php)
Added default rss feed to all pages.

Set the posts to display showing the Excerpt if there is one. If there isn't an excerpt, they will use the content (until the more tag).

Created page for 404 responses. It just displays a message and shows a search form that takes you to a page based on index.php