Removed code pertaining to the previous layout system as we have changed the way we are laying out posts.
postContent will handle everything related to the layout of a post. All it needs to be is in the loop (it won't work if it isn't).

homepage.php is a page template that generates a custom loop that loops through a specific category (ie the category with front page posts.

index.php is the "posts" page - it is necessary for wordpress