Worked on the schedule page.
To make it easier to add a show, the page.php template decides if the page being viewed is a child of the overall schedule page. If it is, it renders it with show.php. Otherwise, it renders it with normalPage.php

Right now, show.php just renders the show the same way as normalPage.php, but we are very likely to eventually make it render shows differently.

We sort the shows we get by day of week and time, using the comparater in functions.php