Worked on making the "Show" post type as opposed to use "Pages" with a specific parent. This greatly increases usability.
Fixed bug in sting_compare_shows_by_date_time (it was missing Saturday)

Sped up sting_display_show_schedule by passing in the number of pages (to loop to) as opposed to counting the number of child pages over and over again.

Put the show type into a variable ($show_type) - the type is 'sting_shows,' but we rewrite the url so that it is /shows/

Shows use the microphone icon in the backend.

Added to header the ability to have a default page banner (subTemplates/banners/banner.php) and the one for the homepage (subTemplates/banners/banner-home.php

Moved banners into those templates.

Allowed us to specify the query the loop loops through. If $sting_query is defined, loop.php will loop through $sting_query. Otherwise, it will loop through the default, $wp_query. This allows us to have dj posts on their own page as the main query for that page will give the loop only that page.

Page title is now handled in the banners, not normalPage.php

Modified schedule.php to account for the $show_type and the not recounting the number of pages in the loop.

Removed show.php. It is now single-sting_shows.php

single.php now checks if comments are enabled (and does a better job of displaying next/previous links.