Lots of changes. I should commit more often. There are too many unrelated changes.
Moved js libraries to /js/lib/ folder

404.php - page titles moved up into the page banner (banner.php)
app.css - new app.css with css for the media element we are using for playing the stream
footer.php - now has the player. toggleStream() can be found in stream.js.

Stream.js:
If the stream is playing, stop it. If it is not playing, start it. When the page loads, if the stream was playing on the previous page (using cookies which last for the session), keep it playing.

Functions.php - uses local copy of the Raleway font, added mediaelement styles and the stream.js. Also adds listen-count/livecounter.js ONLY if it is on a page that has the listener counter.

CSS in header needs to be moved to app.css

onAirView.php - has both the chat and the listener count

Moved number of posts to get in one query from the schedule.php to a setting (see admin/admin.php)
Added a search page that will work once we figure out post layout
Added the homepage slider to the home banner (banners/banner-home.php)

Moved page titles to banner.php

Created options for which category's posts should be displayed on the homepage, how many shows to get in one request and which slider/gallery should be used on the homepage.