Resynced with the actual server.
Commented out error log printouts as it was getting too cluttered.
We now send the url to the current show with the rest of the show information so we can link to the show page

Renamed boxes and added scripts to the admin page. Also fixed result processing

Moved the key we use to store the google calendar event id to a variable from being hardcoded in

Removedthe alert from datetime.js and set the datetime format.

removed the archive page /sting/shows, but the page exists, just gives a blank page. Want it to redirect to schedule but might do that with a plugin

got the admin message working (aka this site is going down for updates)

got the fadein/fadeout on the now-playing bar working (see app.js)

When a show isn't playing, it doesn't show the current song because it would be inaccurate

formatting with regard to the now playing menu

single-sting_shows.php should now allow 1 dj to have multiple shows

worked on page titles in banner.php