Worked on the show archive feature (/shows/)
showArchive.php currently has two versions of the show archive page, one is a table like setup and the other uses foundation's columns. We currently prefer the one that uses the columns - the table like one will be removed before the next commit.

Currently it only sorts by first letter of the show's title. The plan is to expand it to allow for sorting by the name of the djs (and have the shows appear multiple times if it has more than one dj).

app.css - see app.scss

functions.php - added comparator for shows by title (using strcmp)

loop.php - added the date posted to each element.

app.scss - fixed half of the mobile bug (the horizontal scroll that shouldn't be there). Formatted the post date and adjusted the posts on the show pages. Code for the show-archive page.