Added documentation and comments throughout the site.
Simplified loop.php by removing the code that prepares the data to be added to the array when we originally were going to send all the data to the client via json. We never did, and one of the lines (the array_push) line helped cause our out of memory error earlier this year.

Commented out most of page.php, going to bring the content from normalPage.php into page.php in the next commit, as well as remove normalPage.php

New screenshot.jpg

Renamed showArchive.php archive-sting_shows.php. Also brought in the use of foundation's equalizer to make all the boxes the same height (foundation.equalizer.js)

Added twitter and facebook buttons to show pages right below the description but above the newline.

Also use the new version of querying for the coauthors of the posts we want based on https://vip.wordpress.com/documentation/list-posts-by-a-co-author/
Also made these same changes to rss-show.php

Made sure that single.php worked with coauthors

Fixed a bug in banner.php that was the source of the problem with /2015/ and /2015/02 type pages - the wp_locale object needed to be included as a global variable.