Fixed listener-count.php due to the fact that we are not allowed to use file_get_contents on IT's servers.
Changed how we handle break schedule to allow shows to still show up when we are on a break.
Also allowed a header image for the music blog as well as a category select for it.

Changed the way we handle show start and end times - we now store them as HH:MM and use strtotime to convert to unix timestamps. This came from an update in Advanced Custom Fields which broke our ability to store them as timestamps. Affected calendar-push.php, functions.php (sorting and displaying)

Hopefully set up the code on google_auth2.js to allow us to do oauth properly. Moved the account registered and the calendar to wrur.thesting@gmail.com

Show Archive and Calendar page edited to let viewers know that we are on a break schedule.

Some appearance changes (see app.scss)

Removed the slider pending future update (on local dev server, hasn't been moved on from there)

Added show archive system.